Since working on a new DirectX driver (which is getting there, I think), I've made some minor documentation updates to .h files. So following on from Pizza, here they are, as well as some Win32 compiler warning fixes (and I actually used the correct types this time!!). There shouldn't be any problems but I haven't been able to test on Linux yet (wrong autoconf version) so please let me know if there are.
Note that in src/gfx/operations.c, I removed quite a lot of 1.0 * float_num. Was multiplying these numbers by 1.0 meant to mean something, such as type casting?


2002-09-15  Alexander R Angas  <wgd@internode.on.net>

	* src/include/win32/sci_win32.h: New file (contains sleep() and
	  other Win32 macros).

	* src/sound/pcmout.c: Added NO_PCMOUT compiler directive if ever required.

	* src/gfx/drivers/ggi_driver.c: Removed unnecessary if statement.

	* src/engine/kmovement.c:
	* src/engine/kstring.c:
	* src/engine/scriptdebug.c:
	* src/engine/vm.c:
	* src/gfx/operations.c:
	* src/gfx/resource/sci_pic_0.c:
	* src/scicore/tools.c:
	* src/scicore/vocab_debug.c:
	  Compiler warning fixes on Win32.

	* src/include/engine.h:
	* src/scicore/tools.c:
	* src/sound/buff_ss.c:
	* src/sound/buff_ss_win32.c:
	* src/sound/event_ss_win32.c:
	* src/sound/midiout_win32mci.c:
	* src/sound/polled_ss_sdl.c:
	* src/win32/fsci_dll.dsp:
	* src/win32/sciv.dsp:
	  Updates for move of Win32 sleep().

	* src/gfx/drivers/sdl_driver.c:
	* src/include/kernel.h:
	* src/include/gfx_driver.h:
	* src/include/gfx_operations.h:
	* src/include/gfx_tools.h:
	* src/include/scitypes.h:
	* src/include/script.h:
	  Minor documentation updates.

+++ src/engine/kmovement.c
@@ -56,7 +56,7 @@ kSetJump(state_t *s, int funct_nr, int argc, heap_ptr argp)
 	SCIkdebug(SCIkBRESEN, "t1: %d, t2: %d\n", t1, t2);
 
 	if (t2) {
-		x = sqrt(abs((gy * dx * dx) / (2 * t2)));
+		x = (int)sqrt(abs((gy * dx * dx) / (2 * t2)));
 		if (t2 *dx < 0)
 			x = -x;
 	}
@@ -64,7 +64,7 @@ kSetJump(state_t *s, int funct_nr, int argc, heap_ptr argp)
 	if (dx >= 0)
 		y=-y;
 	if ((dy < 0) && (!y))
-		y = -sqrt(-2 * gy * dy);
+		y = -(int)sqrt(-2 * gy * dy);
 
 	SCIkdebug(SCIkBRESEN, "SetJump for object at %x", object);
 	SCIkdebug(SCIkBRESEN, "xStep: %d, yStep: %d\n", x, y);
